Convert float -> double for spart.sf_data.diffusion_matrix_K

+++ src/star_formation/GEAR/star_formation_struct.h
@@ -59,7 +59,7 @@ struct star_formation_spart_data {
   long long progenitor_id;
 
 #if defined(CHEMISTRY_GEAR_MF_DIFFUSION) || defined(CHEMISTRY_GEAR_MF_HYPERBOLIC_DIFFUSION)
-  float norm_matrix_K;
+  double norm_matrix_K;
 #endif
 
 };
